refactor: architecture, schema
