+++ prisma/schema.prisma
@@ -9,3 +9,26 @@ datasource db {
   provider = "mysql"
   url      = env("DATABASE_URL")
 }
+
+model User {
+  id        Int      @id @default(autoincrement())
+  email     String   @unique
+  password  String
+}
+
+model Profile {
+  id        Int      @id @default(autoincrement())
+  name      String
+  url       String   @unique
+  description String
+  tabs      Tab[]
+}
+
+model Tab {
+  id       Int      @id @default(autoincrement())
+  title    String
+  type     String
+  info     Json
+  profile  Profile  @relation(fields: [profileId], references: [id])
+  profileId Int
+}
